Update Login Page with Sign Up Button (#6)
* Added:basic login page (but you can't login yet)

* Update App.js

Added some styles, made all the font in the center of the screen. Changed to color of all text to blue.

* Login page are now in a separate file.

I made the login page into a separate file. I also create another file called Reg.js, it is for the registration page and it is empty.

* New register button

Added a new register button for register accounts. However, it currently has no affects.

* Added space: Login.js

Added space to make Login.js

Co-authored-by: lybned <60757167+lybned@users.noreply.github.com>